feat: specify Ollama host and port in opts
This facilitates the case where the Ollama server is running on a
different computer from the one running the neovim instance. By adding
host and port options, the user can set those appropriately to point to
their Ollama server without needing to modify the `command` and
`list_models` table values.