feat: support changes to nvim_buf_set_text()
Current nightly versions of neovim have changed the behaviour of the
cursor updating in the nvim_buf_set_text() api function. As a result,
when using gen.nvim in split mode, the buffer no longer scrolls as
expected when ollama generates text that scrolls past the end of the
viewable area. This PR adds support for proper buffer scrolling in
nightly neovim versions (0.10-dev) while maintaining the functionality
in the current stable versions of neovim.