fix: parse opts.prompt instead of prompt
An earlier commit (ed802f06a691895c8d6c5cf88b70b698a52514ad) introduced
the `run_command` function. As a result of this function, the local
variable `prompt` does not exist in the scope of the `run_command`
function. Instead, when `opts.show_prompt` is true, we now parse the
`opts.prompt` variable, which is available in the `run_command`
function.